Refactor code
1. Remove exportJarFileCommand.ts. The `ExportJarStep` and `stepMap` are moved to `utility.ts` and the functions are moved to `ExportJarTaskProvider.ts` as a new class called ExportJarCommand
