+++ src/exportJarSteps/GenerateJarExecutor.ts
@@ -7,11 +7,10 @@ import * as _ from "lodash";
 import { basename, dirname, extname, isAbsolute, join, normalize, relative } from "path";
 import { Disposable, ProgressLocation, QuickInputButtons, QuickPickItem, Uri, window } from "vscode";
 import { sendInfo } from "vscode-extension-telemetry-wrapper";
-import { ExportJarStep } from "../exportJarFileCommand";
 import { Jdtls } from "../java/jdtls";
 import { IExportJarStepExecutor } from "./IExportJarStepExecutor";
 import { IClasspath, IStepMetadata } from "./IStepMetadata";
-import { createPickBox, ExportJarTargets, getExtensionApi,
+import { createPickBox, ExportJarStep, ExportJarTargets, getExtensionApi,
     resetStepMetadata, saveDialog, toPosixPath } from "./utility";
 
 export class GenerateJarExecutor implements IExportJarStepExecutor {
