+++ src/exportJarSteps/IExportJarStepExecutor.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { ExportJarStep } from "../exportJarFileCommand";
 import { IStepMetadata } from "./IStepMetadata";
+import { ExportJarStep } from "./utility";
 
 export interface IExportJarStepExecutor {
     getNextStep(): ExportJarStep;
