+++ src/exportJarSteps/ResolveJavaProjectExecutor.ts
@@ -3,13 +3,12 @@
 
 import * as _ from "lodash";
 import { Disposable, QuickPickItem, Uri, workspace, WorkspaceFolder } from "vscode";
-import { ExportJarStep } from "../exportJarFileCommand";
 import { Jdtls } from "../java/jdtls";
 import { INodeData } from "../java/nodeData";
 import { WorkspaceNode } from "../views/workspaceNode";
 import { IExportJarStepExecutor } from "./IExportJarStepExecutor";
 import { IStepMetadata } from "./IStepMetadata";
-import { createPickBox } from "./utility";
+import { createPickBox, ExportJarStep } from "./utility";
 
 export class ResolveJavaProjectExecutor implements IExportJarStepExecutor {
